pyproject.toml
[tool.poetry]
name = "cloudevents_python"
version = "0.1.0"
description = ""
authors = ["ayekhantkhantpo <ayekhantkhantpo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.21.0"
Flask = "^2.2.2"
cloudevents = "^1.9.0"
requests = "^2.28.2"
kafka-python = "^2.0.2"
structlog = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
isort = "^5.11.4"
black = "^22.12.0"
pyproject-flake8 = "^5.0.4"

[tool.poetry.scripts]
start_server = "cloudevents_python.sample_server:main"
with_requests = "cloudevents_python.with_requests:main"
with_kafka = "cloudevents_python.with_kafka:main"
kafka_consumer = "cloudevents_python.kafka_consumer:main"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.flake8]
max-line-length = 88
max-complexity = 18
ignore = "E203, E266, E501, E503"
select = "B, C, E, F, W, T4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
